gives score recalculation jobs more accurate name, also adds a new job to run a score recalculation on all active courses
