+++ app/models/course.rb
@@ -216,6 +216,12 @@ class Course < ApplicationRecord
     end
   end
 
+  def recalculate_active_student_scores
+    ordered_active_student_ids.each do |student_id|
+      ScoreRecalculatorJob.perform_async(student_id, self.id)
+    end
+  end
+
   def formatted_long_name
     if semester.present? && year.present?
       "#{self.course_number} #{self.name} #{(self.semester).capitalize} #{self.year}"
@@ -236,6 +242,15 @@ class Course < ApplicationRecord
     User
       .joins(:course_memberships)
       .where("course_memberships.course_id = ? and course_memberships.role = ?", self.id, "student")
+      .select(:id) # only need the ids, please
+      .order("id ASC")
+      .collect(&:id)
+  end
+
+  def ordered_active_student_ids
+    User
+      .joins(:course_memberships)
+      .where("course_memberships.course_id = ? and course_memberships.role = ? and course_memberships.active = ?", self.id, "student", true)
       .select(:id) # only need the ids, please
       .order("id ASC")
       .collect(&:id)
