+++ config/sidekiq.yml
@@ -5,11 +5,18 @@
     description: "This job triggers an export of all courses created within the \
                   last month and emails it to a list of recipients as a CSV attachment."
 
-  recalculate_active_course_scores:
+  recalculate_recently_updated_scores:
     cron: "0 2 * * *"
-    class: RecalculateActiveCourseScoresJob
+    class: RecalculateRecentlyUpdatedScoresJob
     description: "This job triggers a score recalculation job for every student \
-                  belonging to an active course."
+                  that has had a grade updated in the last 24 hours, this includes \
+                  grades, earned badges, and team challenges."
+
+  recalculate_active_course_scores:
+    cron: "0 3 * * SAT"
+    class: RecalculateActiveCourseScoresJob
+    description: "This job triggers a score recalculation job for every \
+                  active student in every active course."
 
 :queues:
   - high
